+++ Plugin/LegacyPatches/System/IO/DirectoryPatch.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Reflection;
 using System.Security.AccessControl;
 using HarmonyLib;
-using UniTASPlugin.LegacyFakeGameState.GameFileSystem;
+using UniTASPlugin.GameEnvironment.InnerState.FileSystem.OsFileSystems;
 using UniTASPlugin.ReverseInvoker;
 using DirOrig = System.IO.Directory;
 using FileOrig = System.IO.File;
