+++ Plugin/LegacyPatches/System/IO/PathPatch.cs
@@ -3,7 +3,7 @@ using System.Globalization;
 using System.Reflection;
 using System.Text;
 using HarmonyLib;
-using UniTASPlugin.LegacyFakeGameState.GameFileSystem;
+using UniTASPlugin.GameEnvironment.InnerState.FileSystem.OsFileSystems;
 using UniTASPlugin.ReverseInvoker;
 using DirOrig = System.IO.Directory;
 using PathOrig = System.IO.Path;
